% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{alpaca.getMember}
\alias{alpaca.getMember}
\title{get the member vector from alpaca object}
\usage{
alpaca.getMember(alp, target = "all")
}
\arguments{
\item{alp}{alpaca object}

\item{target}{tf, gene, or all}
}
\value{
member vector
}
\description{
get the member vector from alpaca object
}
