% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm.R
\name{annot_at_risk}
\alias{annot_at_risk}
\title{Annotate Plot with Numbers at Risk Table}
\usage{
annot_at_risk(
  gg_plt,
  fit_km,
  font_size = 10,
  annot_at_risk_title = TRUE,
  rel_height_plot = 0.75,
  xlab = "Days"
)
}
\arguments{
\item{gg_plt}{A \code{ggplot2} or \code{cowplot} object of the Kaplan-Meier plot.}

\item{fit_km}{A fitted Kaplan-Meier object of class \code{survfit}, used to generate the table data.}

\item{font_size}{Numeric, base font size for the table.}

\item{annot_at_risk_title}{Logical, whether to include the title "Patients at Risk:".}

\item{rel_height_plot}{Numeric, relative height of the main plot area compared to the 'at-risk' table (0 to 1).}

\item{xlab}{Character string for the x-axis label on the 'at-risk' table (typically time).}
}
\value{
A \code{cowplot} object combining the KM plot and the 'Numbers at Risk' table.
}
\description{
Adds a "Numbers at Risk" table below a Kaplan-Meier plot using \code{cowplot::plot_grid}.
}
