% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm.R
\name{annot_cox_ph}
\alias{annot_cox_ph}
\title{Annotate Kaplan-Meier Plot with Cox-PH Table}
\usage{
annot_cox_ph(
  gg_plt,
  coxph_tbl,
  control_annot_coxph = control_coxph_annot(),
  font_size = 10
)
}
\arguments{
\item{gg_plt}{A \code{ggplot2} or \code{cowplot} object of the Kaplan-Meier plot.}

\item{coxph_tbl}{A data frame containing pre-calculated Cox-PH results,
typically generated by \code{\link{h_tbl_coxph_pairwise}}.}

\item{control_annot_coxph}{A list of control parameters for the annotation box,
typically generated by \code{\link{control_coxph_annot}}.}

\item{font_size}{Numeric, base font size for the annotation table.}
}
\value{
A \code{cowplot} object with the Cox-PH table annotation added.
}
\description{
Adds a Cox Proportional Hazards summary table as an annotation box on a
Kaplan-Meier plot using \code{cowplot}.
}
