% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm.R
\name{annot_surv_med}
\alias{annot_surv_med}
\title{Annotate Kaplan-Meier Plot with Median Survival Table}
\usage{
annot_surv_med(
  gg_plt,
  fit_km,
  control_annot_surv_med = control_surv_med_annot(),
  font_size = 10
)
}
\arguments{
\item{gg_plt}{A \code{ggplot2} or \code{cowplot} object of the Kaplan-Meier plot.}

\item{fit_km}{A fitted Kaplan-Meier object of class \code{survfit}, used to generate the table data.}

\item{control_annot_surv_med}{A list of control parameters for the annotation box.}

\item{font_size}{Numeric, base font size for the annotation table.}
}
\value{
A \code{cowplot} object with the median survival table annotation added.
}
\description{
Adds a median survival time summary table as an annotation box on a
Kaplan-Meier plot using \code{cowplot}.
}
