% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gg_pkc.R
\name{annotate_gg_pkc}
\alias{annotate_gg_pkc}
\title{Annotate PK Plot with Summary Table}
\usage{
annotate_gg_pkc(
  data,
  gg_plt,
  time_var = NULL,
  analyte_var = NULL,
  group = NULL,
  summary_stats = c("n", "mean", "sd"),
  text_size = 3.5,
  rel_height_plot = 0.75,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
The raw or summarized dataset used to generate the table.}

\item{gg_plt}{(\code{ggplot2})\cr
The PK plot.}

\item{time_var, analyte_var, group}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Optional. If \code{NULL} (default), the function automatically extracts these
from the \code{gg_plt} mapping.}

\item{summary_stats}{(\code{character})\cr
A vector of statistics to include. Defaults to \code{c("n", "mean", "sd")}.}

\item{text_size}{(\code{numeric})\cr
The font size for the table text. Defaults to \code{3}.}

\item{rel_height_plot}{(\code{numeric})\cr
Relative height of the plot vs the table. Defaults to \code{0.75}.}

\item{...}{Additional args passed to \code{df2gg_2()}.}
}
\description{
Annotate PK Plot with Summary Table
}
