% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ggplot.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Convert gt/gtsummary table to ggplot}
\usage{
as_ggplot(x, ...)
}
\arguments{
\item{x}{gt or gtsummary table}

\item{...}{arguments passed to \code{gt::gtsave()}}
}
\value{
a ggplot object
}
\description{
useful when you want to place a ggplot and gt table side-by-side.
To use this function you must install the magick R package AND system program
(see https://docs.ropensci.org/magick/articles/intro.html#installing-magick-1)
}
\concept{gtsummary-related functions}
