% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lineplot.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate Summary Statistics for a Vector}
\usage{
calc_stats(x, conf_level = 0.95, decimal_places = 2)
}
\arguments{
\item{x}{(\code{numeric})\cr vector to summarize.}

\item{conf_level}{(\code{probability})\cr The confidence level for the interval (default is 0.95).}

\item{decimal_places}{(\code{integer})\cr Number of decimal places for numeric values in the table.}
}
\value{
A \code{list} with statistics: \code{n}, \code{mean}, \code{mean_ci_lwr}, \code{mean_ci_upr}, \code{median}, \code{sd}.
}
\description{
Calculates N, mean, standard deviation, and a two-sided confidence
interval for the mean of a numeric vector.
}
\examples{
x <- c(10, 20, 40, NA, 30)
calc_stats(x, conf_level = 0.95, decimal_places = 2)
}
