% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_col.R
\name{calculate_widths}
\alias{calculate_widths}
\title{Calculate information for width}
\usage{
calculate_widths(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
A gtsummary table}
}
\value{
A gtsummary table.
}
\description{
Modifies the table
}
\examples{
tbl <- gtsummary::tbl_summary(data = ADSL, by = "SEX", include = "ACTARM") |>
    gtsummary::add_stat_label(location = "column")
calculate_widths(tbl)
}
