% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{condor.createObject}
\alias{condor.createObject}
\title{creates condor object}
\usage{
condor.createObject(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
condor object
}
\description{
creates condor object
}
