% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{condor.run}
\alias{condor.run}
\title{Run CONDOR clustering}
\usage{
condor.run(elist, qscore = F)
}
\arguments{
\item{elist}{edge list}

\item{qscore}{TRUE = output qscore / FALSE = do not output qscore}
}
\value{
condor object
}
\description{
Run CONDOR clustering
}
