% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm.R
\name{control_surv_timepoint}
\alias{control_surv_timepoint}
\title{Control parameters for Survival Timepoint Estimation}
\usage{
control_surv_timepoint(
  conf_level = 0.95,
  conf_type = c("plain", "log", "log-log")
)
}
\arguments{
\item{conf_level}{A numeric value (0 to 1) for the confidence level.}

\item{conf_type}{A character string specifying the type of confidence interval.
Must be one of \code{"plain"}, \code{"log"}, or \code{"log-log"}.}
}
\value{
A list with elements \code{conf_level} and \code{conf_type}.
}
\description{
Creates a list of control parameters for \code{survival::survfit} when used
for timepoint estimation.
}
\keyword{internal}
