% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane.unipartite.R
\name{crane.unipartite}
\alias{crane.unipartite}
\title{Pertrubs the unipartite network with fixed node strength from adjacency matrix}
\usage{
crane.unipartite(A, alpha = 0.1)
}
\arguments{
\item{A}{Adjacency Matrix}

\item{alpha}{alpha paramter perturbs each edge weights}
}
\value{
adjacency matrix
}
\description{
Pertrubs the unipartite network with fixed node strength from adjacency matrix
}
