% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_pkparam_sort.R
\name{d_pkparam}
\alias{d_pkparam}
\title{Generate PK reference dataset}
\usage{
d_pkparam()
}
\value{
A \code{data.frame} of PK parameters.
}
\description{
Generate PK reference dataset
}
\examples{
pk_reference_dataset <- d_pkparam()

}
