% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{elist.addTags}
\alias{elist.addTags}
\title{Adds "_A" to first column and "_B" to second column}
\usage{
elist.addTags(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
edge list
}
\description{
Adds "_A" to first column and "_B" to second column
}
