% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{elist.isEdgeOrderEqual}
\alias{elist.isEdgeOrderEqual}
\title{check if first two columns are identical}
\usage{
elist.isEdgeOrderEqual(elist1, elist2)
}
\arguments{
\item{elist1}{edge list}

\item{elist2}{edge list}
}
\value{
boolean
}
\description{
check if first two columns are identical
}
