% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{elist.sort}
\alias{elist.sort}
\title{Sorts the edge list based on first two columns in alphabetical order}
\usage{
elist.sort(elist)
}
\arguments{
\item{elist}{edge list}
}
\value{
edge list
}
\description{
Sorts the edge list based on first two columns in alphabetical order
}
