% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{fmt_3sig}
\alias{fmt_3sig}
\title{Format to 3 Significant Figures}
\usage{
fmt_3sig(x)
}
\arguments{
\item{x}{(\code{numeric(1)})\cr a single numeric value.}
}
\value{
A \code{character} string, \code{"NE"} if \code{x} is \code{NA}, or \code{NA_character_} if \code{x} is non-finite.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formats a numeric value to 3 significant figures, removing trailing decimals.
}
\examples{
fmt_3sig(0.001234)
fmt_3sig(123456)

}
