% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format to 1 Decimal Place}
\usage{
fmt_pct(x)
}
\arguments{
\item{x}{(\code{numeric(1)})\cr a single numeric value.}
}
\value{
A \code{character} string, \code{"NE"} if \code{x} is \code{NA}, or \code{NA_character_} if \code{x} is non-finite.
}
\description{
Format to 1 Decimal Place
}
\examples{
fmt_pct(45.678)
fmt_pct(NA)

}
