% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{geom_cv}
\alias{geom_cv}
\title{Geometric Coefficient of Variation}
\usage{
geom_cv(x)
}
\arguments{
\item{x}{(\code{numeric})\cr numeric vector of positive values.}
}
\value{
A single numeric value representing the geometric coefficient of variation (\%).
}
\description{
Geometric Coefficient of Variation
}
\examples{
geom_cv(c(1, 2, 3, 4, 5))

}
