% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric Mean}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{(\code{numeric})\cr numeric vector. Non-positive values are ignored.}

\item{na.rm}{(\code{logical(1)})\cr whether to remove \code{NA} values. Default is \code{TRUE}.}
}
\value{
A single numeric value, or \code{NA_real_} if there are no usable values
or if any non-\code{NA} value is <= 0.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
geom_mean(c(1, 2, 3, 4, 5))

}
