% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBearerToken.R
\name{getBearerToken}
\alias{getBearerToken}
\title{Retrieve Bearer Token}
\usage{
getBearerToken(url, username, password, clientId, clientSecret, realm, ...)
}
\arguments{
\item{url}{url}

\item{username}{username}

\item{password}{password}

\item{clientId}{client id}

\item{clientSecret}{client secret}

\item{realm}{realm}

\item{...}{further arguments to \code{\link{POST}}}
}
\description{
Retrieve Bearer Token
}
