% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_roche.R
\name{label_roche}
\alias{label_roche}
\alias{roche_pvalue}
\alias{label_roche_pvalue}
\alias{roche_percent}
\alias{label_roche_percent}
\title{Formatting percent and p-values}
\usage{
roche_pvalue(
  x,
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  decimal.mark = getOption("OutDec"),
  ...
)

label_roche_pvalue(
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  decimal.mark = getOption("OutDec"),
  ...
)

roche_percent(
  x,
  prefix = "",
  suffix = "",
  digits = 1,
  scale = 100,
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  decimal.mark = getOption("OutDec"),
  ...
)

label_roche_percent(
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  suffix = "",
  scale = 100,
  digits = 1,
  decimal.mark = getOption("OutDec"),
  ...
)
}
\arguments{
\item{x}{(\code{numeric})\cr
Numeric vector of p-values.}

\item{big.mark}{(\code{string})\cr
Character used between every 3 digits to separate
hundreds/thousands/millions/etc.
Default is \code{","}, except when \code{decimal.mark = ","} when the default is a space.}

\item{decimal.mark}{(\code{string})\cr
The character to be used to indicate the numeric decimal point.
Default is \code{"."}  or \code{getOption("OutDec")}}

\item{...}{Arguments passed on to \code{base::format()}}

\item{digits}{(\code{integer})\cr
Number of digits large p-values are rounded. Must be 1, 2, or 3.
Default is 1.}
}
\value{
A character vector of rounded p-values
}
\description{
\itemize{
\item \code{label_roche_pvalue()} returns formatted pvalues.
\item \code{label_roche_percent()} returns formatted percent values. This function only formats percentages between 0 and 1.
}
}
\examples{
p-value formatting
x <- c(0.0000001, 0.123456)

roche_pvalue(x)
label_roche_pvalue()(x)

percent formatting
x <- c(0.0008, 0.9998)

roche_percent(x)
label_roche_percent()(x)
}
