% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_zero_recode.R
\name{modify_zero_recode}
\alias{modify_zero_recode}
\title{Zero Count Recode}
\usage{
modify_zero_recode(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
a gtsummary table}
}
\value{
a gtsummary table
}
\description{
This function removes the percentage from cells with zero counts.
For example,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{0 (0.0\%)      -->  0
0 (0\%)        -->  0
0 (NA\%)       -->  0
0 / nn (0\%)   -->  0 / nn
0/nn (0.0\%)   -->  0/nn
0 / 0 (NA\%)   -->  0 / 0
}\if{html}{\out{</div>}}
}
\details{
The function is a wrapper for \code{gtsummary::modify_post_fmt_fun()}.
}
\examples{
trial |>
  dplyr::mutate(trt = factor(trt, levels = c("Drug A", "Drug B", "Drug C"))) |>
  tbl_summary(include = trt) |>
  modify_zero_recode()
}
