% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_md.R
\name{strip_md}
\alias{strip_md}
\alias{strip_md_bold}
\alias{strip_md_italic}
\title{Strip Markdown Bold or Italic}
\usage{
strip_md_bold(x, type = "star")

strip_md_italic(x, type = "star")
}
\arguments{
\item{x}{(\code{character})\cr
A character vector.}

\item{type}{(\code{character})\cr
Specified which type of markdown syntax to strip.
Must be one or both of \code{c("star", "underscore")}}
}
\value{
character vector
}
\description{
Strip Markdown Bold or Italic
}
\examples{
strip_md_bold(c("**Placebo**  \nN=45", "**High Dose**  \nN=54"))
}
