% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_rates_by_grade.R
\name{tbl_ae_rates_by_grade}
\alias{tbl_ae_rates_by_grade}
\alias{add_overall.tbl_ae_rates_by_grade}
\title{AE Rates by Highest Toxicity Grade}
\usage{
tbl_ae_rates_by_grade(
  data,
  grade,
  ae,
  soc,
  denominator,
  by = NULL,
  id = "USUBJID",
  include_overall = all_of(c(soc, ae)),
  statistic = everything() ~ "{n} ({p}\%)",
  label = NULL,
  digits = NULL,
  sort = "descending",
  filter = NULL,
  grade_groups = list(),
  grades_exclude = NULL
)

\method{add_overall}{tbl_ae_rates_by_grade}(
  x,
  last = FALSE,
  col_label = "**Overall**  \\nN = {style_number(N)}",
  statistic = NULL,
  digits = NULL,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{grade}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column name with the toxicity grade levels.}

\item{ae}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column name with the adverse event terms.}

\item{soc}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column name with the system organ class.}

\item{denominator}{(\code{data.frame}, \code{integer})\cr
used to define the denominator and enhance the output.
The argument is required for \code{tbl_hierarchical()} and optional for \code{tbl_hierarchical_count()}.
The \code{denominator} argument must be specified when \code{id} is used to calculate event rates.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
a single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
argument used to subset \code{data} to identify rows in \code{data} to calculate
event rates in \code{tbl_hierarchical()}.}

\item{include_overall}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables from \code{c(soc, ae)} for which an overall section at that hierarchy level should be computed.
An overall section at the \code{soc} variable level will have label \code{"- Any adverse events -"}. An overall section at
the \code{ae} variable level will have label \code{"- Overall -"}. The default is \code{c(soc, ae)}.}

\item{statistic}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
used to specify the summary statistics to display for all variables in \code{tbl_hierarchical()}.
The default is \code{everything() ~ "{n} ({p})"}.}

\item{label}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
used to override default labels in hierarchical table, e.g. \code{list(AESOC = "System Organ Class")}.
The default for each variable is the column label attribute, \code{attr(., 'label')}.
If no label has been set, the column name is used.}

\item{digits}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified,
default formatting is assigned via \code{label_style_number()} for statistics \code{n} and \code{N}, and
\code{label_style_percent(digits=1)} for statistic \code{p}.}

\item{sort}{(\code{string})\cr
type of sorting to perform. Value must be one of:
\itemize{
\item \code{"alphanumeric"} - at each hierarchy level of the table, rows are ordered alphanumerically (i.e. A to Z)
by label text.
\item \code{"descending"} - at each hierarchy level of the table, count sums are calculated for each row and rows are
sorted in descending order by sum. If \code{sort = "descending"}, the \code{n} statistic is used to calculate row sums if
included in \code{statistic} for all variables, otherwise \code{p} is used. If neither \code{n} nor \code{p} are present in \code{x} for
all variables, an error will occur.
}

Defaults to \code{"descending"}.}

\item{filter}{(\code{expression})\cr
An expression that is used to filter rows of the table. See the Details section below for more information.}

\item{grade_groups}{(\code{list})\cr
A list of formulas each corresponding to a grade group for which rates should be calculated. Grade groups must be
mutually exclusive, i.e. each grade cannot be assigned to more than one grade group. To specify each grade group,
create a formula with a vector of grades included on the left-hand side of the formula and the name of the grade
on the right-hand side, e.g. \code{c("1", "2") ~ "Grade 1-2"}.}

\item{grades_exclude}{(\code{character})\cr
A vector of grades to omit individual rows for when printing the table. These grades will still be used when
computing overall totals and grade group totals. For example, to avoid duplication, if a grade group is defined as
\code{"5" ~ "Grade 5"}, the individual rows corresponding to grade 5 can be excluded by setting \code{grades_exclude = "5"}.}

\item{x}{(\code{tbl_ae_rates_by_grade})\cr
A gtsummary table of class \code{'tbl_ae_rates_by_grade'}.}
}
\value{
a gtsummary table of class \code{"tbl_ae_rates_by_grade"}.
}
\description{
A wrapper function for \code{\link[gtsummary:tbl_hierarchical]{gtsummary::tbl_hierarchical()}} to calculate rates of highest toxicity grades with the options
to add rows for grade groups and additional summary sections at each variable level.

Only the highest grade level recorded for each subject will be analyzed. Prior to running the function, ensure that
the toxicity grade variable (\code{grade}) is a factor variable, with factor levels ordered lowest to highest.

Grades will appear in rows in the order of the factor levels given, with each grade group appearing prior to the
first level in its group.
}
\details{
When using the \code{filter} argument, all grade and grade group rows that meet the given filtering criteria will be kept.
If a grade group row meets the filter criteria but the individual grade rows within the group do not, the grade group
row will still be kept. If all rows in a given table hierarchy section have been filtered out, the summary row of
that section will also be excluded. For example, if \code{soc = AESOC} and no grade or grade group rows in the section
corresponding to \code{AESOC = "CARDIAC DISORDERS"} meet the filtering criteria, the \code{"CARDIAC DISORDERS"} summary row and
all subsequent rows in this section will be removed from the table. Filtering out rows does not exclude the records
corresponding to these rows from being included in rate calculations for overall sections.

See the \code{filter} argument of \code{\link[gtsummary:filter_hierarchical]{gtsummary::filter_hierarchical()}} for more details and examples.
}
\examples{
ADSL <- cards::ADSL |> mutate(TRTA = ARM)
ADAE_subset <- cards::ADAE |>
  dplyr::filter(
    AESOC \%in\% unique(cards::ADAE$AESOC)[1:5],
    AETERM \%in\% unique(cards::ADAE$AETERM)[1:10]
  )

## Add AETOXGR variable to example dataset
set.seed(1)
ADAE_subset <- ADAE_subset |>
  dplyr::rowwise() |>
  mutate(
    AETOXGR = dplyr::case_when(
      AESEV == "MILD" ~ sample(1:2, 1),
      AESEV == "MODERATE" ~ sample(3:4, 1),
      AESEV == "SEVERE" ~ 5,
    ) |> factor(levels = 1:5)
  ) |>
  dplyr::ungroup()

# Example 1 ----------------------------------
grade_groups <- list(
  c("1", "2") ~ "Grade 1-2",
  c("3", "4") ~ "Grade 3-4",
  "5" ~ "Grade 5"
)

tbl_ae_rates_by_grade(
  ADAE_subset,
  grade = AETOXGR,
  ae = AEDECOD,
  soc = AEBODSYS,
  denominator = ADSL,
  by = TRTA,
  label = list(
    AEBODSYS = "MedDRA System Organ Class",
    AEDECOD = "MedDRA Preferred Term",
    AETOXGR = "Grade"
  ),
  grade_groups = grade_groups,
  grades_exclude = "5"
)
}
