% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_baseline_chg.R
\name{tbl_baseline_chg}
\alias{tbl_baseline_chg}
\title{Change from Baseline}
\arguments{
\item{digits}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified,
default formatting is assigned via \code{label_style_number()} for the \code{n} statistic and
\code{label_style_number(digits=1, scale=100)} for the \code{p} statistic.}
}
\value{
a gtsummary table
}
\description{
Computes Change from Baseline
}
\examples{
# Example 1 ----------------------------------
}
