% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_demographics.R
\name{tbl_demographics}
\alias{tbl_demographics}
\title{Demographics Table}
\usage{
tbl_demographics(
  data,
  by = NULL,
  label = NULL,
  statistic = list(all_continuous() ~ c("{mean} ({sd})", "{median} ({p25}, {p75})",
    "{min}, {max}"), all_categorical() ~ "{n} ({p}\%)"),
  digits = NULL,
  type = NULL,
  value = NULL,
  sort = all_categorical(FALSE) ~ "alphanumeric",
  percent = c("column", "row", "cell"),
  include = everything()
)
}
\arguments{
\item{data}{(\code{data.frame})\cr A data frame.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}.}

\item{label}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Used to override default labels in summary table, e.g. \code{list(age = "Age, years")}.
The default for each variable is the column label attribute, \code{attr(., 'label')}.
If no label has been set, the column name is used.}

\item{statistic}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies summary statistics to display for each variable.  The default is
\code{list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}\%)")}.
See below for details.}

\item{digits}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s)
or function(s). If not specified, default formatting is assigned
via \code{assign_summary_digits()}. See below for details.}

\item{type}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the summary type. Accepted value are
\code{c("continuous", "continuous2", "categorical", "dichotomous")}.
If not specified, default type is assigned via
\code{assign_summary_type()}. See below for details.}

\item{value}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies the level of a variable to display on a single row.
The gtsummary type selectors, e.g. \code{all_dichotomous()}, cannot be used
with this argument. Default is \code{NULL}. See below for details.}

\item{sort}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies sorting to perform for categorical variables.
Values must be one of \code{c("alphanumeric", "frequency")}.
Default is \code{all_categorical(FALSE) ~ "alphanumeric"}.}

\item{percent}{(\code{string})\cr
Indicates the type of percentage to return.
Must be one of \code{c("column", "row", "cell")}. Default is \code{"column"}.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in the summary table. Default is \code{everything()}.}
}
\value{
a 'gtsummary' table
}
\description{
This is a thin wrapper of \code{gtsummary::tbl_summary()}.
The function is meant for \code{'continuous2'} and \code{'categorical'} summary types only.
\itemize{
\item Continuous variables default to type \code{'continuous2'}.
\item The number of non-missing is placed on the first row for all summaries.
\item The default formatting function for percentages is \code{gtsummary::label_style_number(digits = 1, scale = 100)}.
\item The \code{gtsummary::add_stat_label()} is run automatically after the function's execution.
}
}
\examples{
theme_gtsummary_roche()
tbl <-
  trial |>
  tbl_demographics(
    by = trt,
    include = c(age, grade)
  )

tbl

# extract ARD from table
gather_ard(tbl)

reset_gtsummary_theme()
}
