% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grade_groups.R
\name{tbl_hierarchical_groups}
\alias{tbl_hierarchical_groups}
\title{AE Analysis of Grade Groups}
\usage{
tbl_hierarchical_groups(
  data,
  variables,
  level_groups = list(),
  levels_exclude = NULL,
  id,
  denominator,
  by = NULL,
  include = everything(),
  statistic = everything() ~ "{n} ({p\%})",
  overall_row = FALSE,
  labels = NULL,
  digits = NULL
)
}
\description{
AE Analysis of Grade Groups
}
