% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_listing.R
\name{tbl_listing_as_overloading}
\alias{tbl_listing_as_overloading}
\alias{as_gt}
\alias{as_flex_table}
\alias{as_gt.tbl_listing}
\alias{as_flex_table.tbl_listing}
\title{\verb{as_*()} overloading for \code{tbl_listing}}
\usage{
as_gt(x, ...)

as_flex_table(x, ...)

\method{as_gt}{tbl_listing}(x, ...)

\method{as_flex_table}{tbl_listing}(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
An object of class \code{"gtsummary"}}

\item{...}{Arguments passed on to \code{gt::gt(...)}}
}
\description{
\verb{as_*()} overloading for \code{tbl_listing}
}
\keyword{internal}
