% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_nested_stack.R
\name{tbl_nested_stack}
\alias{tbl_nested_stack}
\title{Nested Stacking}
\usage{
tbl_nested_stack(tbls, group_header, quiet = TRUE)
}
\arguments{
\item{tbls}{(\code{list})\cr
List of gtsummary objects}

\item{group_header}{(\code{list})\cr
list of character vectors, where the character vectors are the nested headers,
e.g.\code{list(c("Table 1", "Sub-table a"), c("Table 2", "Sub-table b"))} would
result in the first tbl being nested under headers \code{c("Table 1", "Sub-table a")},
and the second tbl nested under \code{c("Table 2", "Sub-table b")}.}

\item{quiet}{(scalar \code{logical})\cr
Logical indicating whether to suppress additional messaging. Default is \code{FALSE}.}
}
\value{
a gtsummary table
}
\description{
Nested Stacking
}
\examples{
tbl <- tbl_summary(trial, include = age)

tbl_nested_stack(
  list(tbl, tbl),
  group_header = list(c("table 1", "table a"),
                      c("table 2", "table b"))
)
}
