% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split_horizontal.R
\name{tbl_split_horizontal}
\alias{tbl_split_horizontal}
\title{Horizontal Table Splitting}
\usage{
tbl_split_horizontal(x, keys = first_column(x), groups)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
a gtsummary table}

\item{keys}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to be repeated in each table split.}

\item{groups}{(list of \code{character} vectors)\cr
list of column names that appear in \code{x$table_body}.
Each group of column names represent}
}
\value{
a gtsummary table
}
\description{
Split a gtsummary into parts horizontally (that is, splits by columns in
\code{x$table_body}), and return a list of split tables.
Run \code{gtsummary::show_header_names()} to print all column names to split by.
}
\examples{
# Example 1 ----------------------------------
tbl <- tbl_demographics(cards::ADSL, by = ARM, include = c(AGE, SEX))

# print column names
gtsummary::show_header_names(tbl)

tbl |>
  tbl_split_horizontal(
    keys = c("label", "stat_1"),
    groups = list("stat_2", "stat_3")
  )
}
