% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forest.R
\name{add_forest}
\alias{add_forest}
\title{Add a Forest Plot Column to a gtsummary Table}
\usage{
add_forest(
  x,
  estimate = starts_with("estimate"),
  conf_low = starts_with("conf.low"),
  conf_high = starts_with("conf.high"),
  pvalue = starts_with("p.value"),
  after = starts_with("p.value"),
  table_engine = c("flextable", "gt")
)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
A gtsummary table with estimates and confidence intervals in the table body. Usually produced
by \code{\link[=tbl_roche_subgroups]{tbl_roche_subgroups()}}.}

\item{estimate}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Estimate column name.}

\item{conf_low}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Confidence interval lower bound column name.}

\item{conf_high}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Confidence interval upper bound column name.}

\item{pvalue}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}, optional)\cr
P-value column name. Point sizes in the forest plot will be scaled
according to p-value (smaller p-values = smaller points). \code{NULL} to turn this off.}

\item{after}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Column name after which the forest plot column will be added. Default is after
the p-value column.}

\item{table_engine}{(\code{character})\cr
Table rendering engine to use. Default is "flextable".}
}
\value{
a gt table or flextable object with an added forest plot column.
}
\description{
This function adds a forest plot column to a gtsummary table, typically produced
by \code{\link[=tbl_roche_subgroups]{tbl_roche_subgroups()}}. The forest plot visualizes estimates and confidence intervals
for each subgroup in the table. The function supports rendering with either the \code{gt}
or \code{flextable} engines, making it suitable for different outputs.
}
\details{
Both gt and flextable outputs could produce issues in line continuity between rows if
there are wrapping in the statistical cells.
}
\examples{
# Simple example ------------------------------------------------------------
trial |>
  select(age, marker, grade, response) |>
  tbl_uvregression(
    y = response,
    method = glm,
    method.args = list(family = binomial),
    exponentiate = TRUE,
    hide_n = TRUE
  ) |>
  modify_column_merge(
    pattern = "{estimate} (95\% CI {ci}; {p.value})",
    rows = !is.na(estimate)
  ) |>
  modify_header(estimate = "**Odds Ratio**") |>
  add_forest(table_engine = "gt")

# Realistic example ---------------------------------------------------------
trial |>
  tbl_roche_subgroups(
    rsp = "response",
    by = "trt",
    subgroups = c("grade", "stage"),
    ~ glm(response ~ trt, data = .x) |>
      gtsummary::tbl_regression(
        show_single_row = trt,
        exponentiate = TRUE
      )
  ) |>
  add_forest(pvalue = starts_with("p.value"), table_engine = "flextable")

}
