% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gg_km.R
\name{annotate_gg_km}
\alias{annotate_gg_km}
\alias{annotate_riskdf}
\alias{annotate_surv_med}
\alias{annotate_coxph}
\title{Annotate Kaplan-Meier Plot}
\usage{
annotate_riskdf(
  gg_plt,
  fit_km,
  title = "Patients at Risk:",
  rel_height_plot = 0.75,
  xlab = "Days",
  ...
)

annotate_surv_med(
  gg_plt,
  surv_tbl,
  table_position = c(x = 0.8, y = 0.85, w = 0.32, h = 0.16),
  ...
)

annotate_coxph(
  gg_plt,
  coxph_tbl,
  table_position = c(x = 0.29, y = 0.51, w = 0.4, h = 0.125),
  ...
)
}
\arguments{
\item{gg_plt}{(\code{ggplot2} or \code{cowplot})\cr
The primary plot object of the Kaplan-Meier plot. Note: While floating
tables accept \code{cowplot} objects, aligned tables (like the risk table)
require a pure \code{ggplot2} object.}

\item{fit_km}{(\code{survfit})\cr
A fitted Kaplan-Meier object of class \code{survfit} (from the \code{survival}
package). This object contains the necessary survival data used by
\code{annotate_riskdf()} to calculate and generate the content displayed.}

\item{title}{(\code{character})\cr
A single string value indicating whether to include a title above the
table. Defaults to \code{"Patients at Risk:"}. If \code{NULL}, no title is added.}

\item{rel_height_plot}{(\code{numeric})\cr
A single numeric value defining the \strong{relative height} of the main
Kaplan-Meier plot area compared to the 'at-risk' table. This value should
be between 0 and 1, where a value closer to 1 gives the main plot more
vertical space. Defaults to \code{0.75}.}

\item{xlab}{(\code{character})\cr
A single character string for the \strong{x-axis label} on the 'at-risk' table.
This typically represents time (e.g., "Days").}

\item{...}{Additional arguments passed to the control list for the annotation
box. These arguments override the default values.\cr
Accepted arguments include:
\itemize{
\item \code{fill} (\code{logical}): Whether the annotation box should
have a background fill. Default is \code{TRUE}.
\item \code{font_size} (\code{numeric}): Base font size for the text
inside the annotation box. Default is \code{10}.
}}

\item{surv_tbl}{(\code{data.frame})\cr
A data frame containing the pre-calculated survival summary results,
such as the output from \code{get_surv_times_df()}.}

\item{table_position}{(\code{numeric})\cr
A named numeric vector \code{c(x, y, w, h)} defining the position and size of
the floating table. \code{x} and \code{y} are the coordinates (0 to 1),
while \code{w} and \code{h} represent width and height (0 to 1). Defaults vary
by function.}

\item{coxph_tbl}{(\code{data.frame})\cr
A data frame containing the pre-calculated Cox-PH results, derived
using function \code{get_cox_pairwise_df()}.}
}
\description{
These functions provide capabilities to annotate Kaplan-Meier plots (\code{\link[=gg_km]{gg_km()}})
with additional summary tables, including median survival times, numbers at
risk, and cox proportional hazards results. The annotations are added using
the \code{cowplot} package for flexible placement.
}
\examples{
\dontshow{if (requireNamespace("survival", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Preparing the Kaplan-Meier Plot
library(survival)
use_lung <- lung
use_lung$arm <- factor(sample(c("A", "B", "C"), nrow(use_lung), replace = TRUE))
use_lung$status <- use_lung$status - 1 # Convert status to 0/1
use_lung <- na.omit(use_lung)

formula <- Surv(time, status) ~ arm
fit_kmg01 <- survfit(formula, use_lung)
surv_plot_data <- process_survfit(fit_kmg01)

plt_kmg01 <- gg_km(surv_plot_data)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gg_km]{gg_km()}}, \code{\link[=process_survfit]{process_survfit()}}, and \code{\link[=get_cox_pairwise_df]{get_cox_pairwise_df()}} for
related functionalities.
}
