% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane-package.R
\docType{package}
\name{crane-package}
\alias{crane}
\alias{crane-package}
\title{crane: Supplements the 'gtsummary' Package for Pharmaceutical Reporting}
\description{
Tables summarizing clinical trial results are often complex and require detailed tailoring prior to submission to a health authority. The 'crane' package supplements the functionality of the 'gtsummary' package for creating these often highly bespoke tables in the pharmaceutical industry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/crane}
  \item \url{https://insightsengineering.github.io/crane/}
  \item Report bugs at \url{https://github.com/insightsengineering/crane/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com} (\href{https://orcid.org/0009-0000-8738-5561}{ORCID})
  \item Davide Garolini \email{davide.garolini@roche.com}
  \item Abinaya Yogasekaram \email{abinaya.yogasekaram@contractors.roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
