% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crane-package.R
\docType{package}
\name{crane-package}
\alias{crane}
\alias{crane-package}
\title{crane: Supplements functions in the 'gtsummary' package}
\description{
We often need to make slight modifications to functions found in the 'gtsummary' package to fit the purposes of trial reporting. This package houses those functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/crane}
  \item \url{https://insightsengineering.github.io/crane/}
  \item Report bugs at \url{https://github.com/insightsengineering/crane/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
