% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x)
}
\arguments{
\item{x}{(\code{numeric})\cr numeric vector.}
}
\value{
A single numeric value representing the coefficient of variation (\%).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
cv(c(1, 2, 3, 4, 5))

}
