% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{tbl_demographics}
\alias{g_lineplot}
\alias{g_lineplot_table}
\alias{preprocess_lineplot_data}
\title{Deprecated functions}
\usage{
tbl_demographics(..., nonmissing = "always")

g_lineplot(...)

g_lineplot_table(...)

preprocess_lineplot_data(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
