% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_utils.R
\name{df2gg_aligned}
\alias{df2gg_aligned}
\title{Create and Stack an X-Axis Aligned Table}
\usage{
df2gg_aligned(
  df,
  gg_plt,
  y_col = NULL,
  y_labels = NULL,
  title = NULL,
  xlab = NULL,
  show_xaxis = FALSE,
  text_size = 3.5,
  label_size = 10,
  rel_height_plot = 0.75
)
}
\arguments{
\item{df}{(\code{data.frame})\cr The summary table to render.}

\item{gg_plt}{(\code{ggplot2})\cr The main plot to align to and stack on top of.}

\item{y_col}{(\code{character} or \code{NULL})\cr The name of the column containing Y-axis
labels. If \code{NULL}, the \code{rownames(df)} are used.}

\item{y_labels}{(\code{character} or \code{expression})\cr Optional. A vector of formatted
labels (e.g., parsed plotmath) to override the raw strings extracted from \code{y_col}.}

\item{title}{(\code{character} or \code{NULL})\cr Optional title for the table.}

\item{xlab}{(\code{character} or \code{NULL})\cr Optional X-axis label for the table.}

\item{show_xaxis}{(\code{logical})\cr Should the table display an X-axis line, ticks,
and numeric text? Useful for risk tables. Defaults to \code{FALSE}.}

\item{text_size}{(\code{numeric})\cr Size of the data values in the table.}

\item{label_size}{(\code{numeric})\cr Size of the axis and title text.}

\item{rel_height_plot}{(\code{numeric})\cr Proportion of vertical space given to the main plot.}
}
\value{
A combined \code{patchwork} object.
}
\description{
Generates a text table from a dataframe, perfectly aligns its
X-axis to match a provided ggplot, and seamlessly stacks them vertically.
}
