% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cox_pairwise_df.R
\name{.estimate_p_value}
\alias{.estimate_p_value}
\title{Estimate p value based on chosen test type
Using {coin} and {survival} packages}
\usage{
.estimate_p_value(formula, data, test)
}
\description{
Estimate p value based on chosen test type
Using {coin} and {survival} packages
}
\keyword{internal}
