% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helper.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format to 1 Decimal Place}
\usage{
fmt_pct(x)
}
\arguments{
\item{x}{(\code{numeric})\cr a numeric vector.}
}
\value{
A \code{character} vector the same length as \code{x}, with \code{NA_character_}
where \code{x} is \code{NA} or non-finite.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formats numeric values to a single decimal place.
}
\examples{
fmt_pct(45.678)
fmt_pct(c(45.678, NA))

}
