% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_zero_recode.R
\name{modify_zero_recode}
\alias{modify_zero_recode}
\title{Zero Count Recode}
\usage{
modify_zero_recode(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
a gtsummary table}
}
\value{
a gtsummary table
}
\description{
This function converts \code{"0 (0.0\%)"} cells in a summary table to \code{"0"}.

It's a simple wrapper for \code{gtsummary::modify_post_fmt_fun()}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{gtsummary::modify_post_fmt_fun(
  x,
  fmt_fun = ~ ifelse(. \%in\% c("0 (0.0\%)", "0 (NA\%)"), "0", .),
  columns = gtsummary::all_stat_cols()
)
}\if{html}{\out{</div>}}
}
\examples{
gtsummary::trial |>
  dplyr::mutate(trt = factor(trt, levels = c("Drug A", "Drug B", "Drug C"))) |>
  gtsummary::tbl_summary(include = trt) |>
  modify_zero_recode()
}
