% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_stat_columns.R
\name{protect_stat_columns}
\alias{protect_stat_columns}
\title{Protect Statistics Columns from DOCX Wrapping}
\usage{
protect_stat_columns(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
A \code{gtsummary} object.}
}
\value{
A modified \code{gtsummary} object.
}
\description{
Replaces standard spaces with non-breaking spaces (\verb{\\u00A0}) dynamically
in all visible statistics columns of a \code{gtsummary} table's body.
This forces the DOCX layout engine to keep statistics (e.g., "12.5 ( 95\%)")
on a single line. Column headers (e.g., cohort names) are completely
protected and remain unaffected.
}
\examples{
library(gtsummary)

tbl <- trial |>
  tbl_summary(by = trt, include = age) |>
  add_p() |>
  protect_stat_columns()

}
