% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{add_overall}
\alias{all_continuous}
\alias{all_categorical}
\alias{all_dichotomous}
\alias{all_stat_cols}
\alias{label_style_number}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{gtsummary}{\code{\link[gtsummary]{add_overall}}, \code{\link[gtsummary:select_helpers]{all_categorical}}, \code{\link[gtsummary:select_helpers]{all_continuous}}, \code{\link[gtsummary:select_helpers]{all_dichotomous}}, \code{\link[gtsummary:select_helpers]{all_stat_cols}}, \code{\link[gtsummary:label_style]{label_style_number}}}
}}

