% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_roche.R
\name{roche_percent}
\alias{roche_percent}
\alias{label_roche_percent}
\title{Round percentages}
\usage{
roche_percent(
  x,
  prefix = "",
  suffix = "",
  digits = 1,
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  decimal.mark = getOption("OutDec"),
  ...
)

label_roche_percent(
  big.mark = ifelse(decimal.mark == ",", " ", ","),
  suffix = "",
  digits = 1,
  decimal.mark = getOption("OutDec"),
  ...
)
}
\arguments{
\item{x}{numeric vector of percentages}

\item{prefix}{(\code{string})\cr
Additional text to display before the number.}

\item{suffix}{(\code{string})\cr
Additional text to display after the number.}

\item{digits}{number of digits to round large percentages (i.e. greater than 10\%).
Smaller percentages are rounded to \code{digits + 1} places.
Default is \code{0}}

\item{big.mark}{(\code{string})\cr
Character used between every 3 digits to separate
hundreds/thousands/millions/etc.
Default is \code{","}, except when \code{decimal.mark = ","} when the default is a space.}

\item{decimal.mark}{(\code{string})\cr
The character to be used to indicate the numeric decimal point.
Default is \code{"."}  or \code{getOption("OutDec")}}

\item{...}{Arguments passed on to \code{base::format()}}
}
\value{
A character vector of rounded percent values
}
\description{
This function only formats percentages between 0 and 1.
}
\examples{
x <- c(0.0008, 0.9998)

roche_percent(x)
label_roche_percent()(x)
}
