% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_roche.R
\name{roche_percent}
\alias{roche_percent}
\title{Round percentages}
\value{
A character vector of rounded percent values
}
\description{
Round percentages
}
\examples{
x <- c(0.0000001, 0.123456)

roche_percent(x)
label_roche_percent()(x)
}
