% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ae_rate_and_count.R
\name{tbl_ae_rate_and_count}
\alias{tbl_ae_rate_and_count}
\alias{add_overall.tbl_ae_rate_and_count}
\title{AE Rates and Counts}
\usage{
tbl_ae_rate_and_count(
  data,
  ae,
  soc,
  denominator,
  hlt = NULL,
  by = NULL,
  id = "USUBJID",
  digits = NULL,
  filter = NULL,
  sort = NULL
)

\method{add_overall}{tbl_ae_rate_and_count}(
  x,
  last = FALSE,
  col_label = "**Overall**  \\nN = {style_number(N)}",
  statistic = NULL,
  digits = NULL,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{ae}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column name with the adverse event terms.}

\item{soc}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column name with the system organ class.}

\item{denominator}{(\code{data.frame}, \code{integer})\cr
used to define the denominator and enhance the output.
The argument is required for \code{tbl_hierarchical()} and optional for \code{tbl_hierarchical_count()}.
The \code{denominator} argument must be specified when \code{id} is used to calculate event rates.}

\item{hlt}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
An \emph{optional} single column name with the higher level terms.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
a single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
argument used to subset \code{data} to identify rows in \code{data} to calculate
event rates in \code{tbl_hierarchical()}.}

\item{digits}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s) or function(s). If not specified,
default formatting is assigned via \code{label_style_number()} for statistics \code{n} and \code{N}, and
\code{label_style_percent(digits=1)} for statistic \code{p}.}

\item{filter, sort}{\emph{Optional} arguments passed to \code{gtsummary::filter_hierarchical(filter)}
and \code{gtsummary::sort_hierarchical(sort)}.
All filtering and sorting is applied to the \emph{rates-related} statistics, and
not the counts.}

\item{x}{(\code{tbl_summary}, \code{tbl_svysummary}, \code{tbl_continuous}, \code{tbl_custom_summary})\cr
A stratified 'gtsummary' table}

\item{last}{(scalar \code{logical})\cr
Logical indicator to display overall column last in table.
Default is \code{FALSE}, which will display overall column first.}

\item{col_label}{(\code{string})\cr
String indicating the column label. Default is \code{"**Overall**  \nN = {style_number(N)}"}}

\item{statistic}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
used to specify the summary statistics to display for all variables in \code{tbl_hierarchical()}.
The default is \code{everything() ~ "{n} ({p})"}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a gtsummary table
}
\description{
A mix of adverse event rates (from \code{tbl_hierarchical()}) and counts
(from \code{tbl_hierarchical_count()}).
}
\examples{
# Example 1 ----------------------------------
cards::ADAE |>
  tbl_ae_rate_and_count(
    denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
    by = TRTA,
    ae = AEDECOD,
    soc = AEBODSYS,
    # Keep rows where AEs across the row have an overall prevalence of greater than 10\%
    filter = sum(n) / sum(N) > 0.10
  ) |>
  add_overall(last = TRUE)
}
