% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_coxph.R
\name{tbl_coxph}
\alias{tbl_coxph}
\title{Pairwise Cox Proportional Hazards Table}
\usage{
tbl_coxph(data, model_formula, arm, ref_group = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr
The dataset containing the variables for the survival model.}

\item{model_formula}{(\code{formula})\cr
A survival formula, e.g., \code{survival::Surv(time, status) ~ arm}.}

\item{arm}{(\code{character})\cr
The name of the treatment arm variable (must be a factor).}

\item{ref_group}{(\code{character})\cr
The reference group level for the treatment arm. Defaults to the first
factor level of the \code{arm} variable.}
}
\value{
A \code{gtsummary} object summarizing the pairwise Cox PH results.
}
\description{
Computes pairwise Cox proportional hazards tests and returns a \code{gtsummary}
object. The table splits the results by comparison arms, presenting the
p-value, Hazard Ratio, and 95\% Confidence Interval in a stacked layout where
statistics form the rows of the table.
}
\examples{
# Setup sample survival data with 3 arms to test pairwise comparisons
surv_data <- survival::lung |>
  dplyr::mutate(
    arm = factor(sample(c("A", "B", "C"), dplyr::n(), replace = TRUE)),
    status = status - 1
  ) |>
  dplyr::filter(dplyr::if_all(dplyr::everything(), ~ !is.na(.)))

# Generate the gtsummary table
tbl_coxph(
  data = surv_data,
  model_formula = survival::Surv(time, status) ~ arm,
  arm = "arm",
  ref_group = "A"
)

# Setup sample survival data with 2 arms to test if table drops comparison
# column
surv_data <- survival::lung |>
  dplyr::mutate(
    arm = factor(sample(c("A", "B"), dplyr::n(), replace = TRUE)),
    status = status - 1
  ) |>
  dplyr::filter(dplyr::if_all(dplyr::everything(), ~ !is.na(.)))
# Generate the gtsummary table
tbl_coxph(
  data = surv_data,
  model_formula = survival::Surv(time, status) ~ arm,
  arm = "arm",
  ref_group = "A"
)

}
