% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_hierarchical_groups.R
\name{tbl_hierarchical_groups}
\alias{tbl_hierarchical_groups}
\title{AE Analysis of Grade Groups}
\usage{
tbl_hierarchical_groups(
  data,
  variables,
  id,
  denominator,
  by = NULL,
  include = dplyr::last(variables),
  include_overall = everything(),
  statistic = everything() ~ "{n} ({p}\%)",
  labels = NULL,
  digits = NULL,
  sort = "alphanumeric",
  filter = NULL,
  level_groups = list(),
  levels_exclude = NULL
)
}
\description{
AE Analysis of Grade Groups
}
\examples{
ADSL <- random.cdisc.data::cadsl
ADAE <- random.cdisc.data::cadae

grade_groups <- list(
  c("1", "2") ~ "Grade 1-2",
  c("3", "4") ~ "Grade 3-4",
  "5" ~ "Grade 5"
)

tbl_hierarchical_groups(
  ADAE |> mutate(AETOXGR = factor(AETOXGR, ordered = TRUE)),
  variables = c(AEBODSYS, AEDECOD, AETOXGR),
  level_groups = grade_groups,
  levels_exclude = "5",
  id = USUBJID,
  denominator = ADSL,
  by = ACTARM,
  labels = list(
    AEBODSYS = "MedDRA System Organ Class",
    AEDECOD = "MedDRA Preferred Term",
    AETOXGR = "Grade"
  )
)
}
