% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_listing.R
\name{tbl_listing}
\alias{tbl_listing}
\title{Create listings from a data frame}
\usage{
tbl_listing(
  data,
  keys = NULL,
  order_by = keys,
  do_not_print_duplicated_keys = TRUE
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame containing the data to be displayed in the listing.}

\item{keys}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to be highlighted on the left of the listing.}

\item{order_by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to be sorted. It defaults to \code{keys}.}

\item{do_not_print_duplicated_keys}{(\code{logical})\cr
whether to add blank values where key columns have duplicate values. Defaults to \code{TRUE}.}
}
\description{
Create listings from a data frame
}
\details{
This function converts a data frame into a listing format and it relies on the user's pre-processing
to decide which column to show and select unique or non-unique rows. Also sorting depends on the user
pre-processing.
}
\note{
\code{NA} values are not printed by default in \code{{gtsummary}}. Please consider making them explicit if
you want them to be displayed in the listing.
}
\examples{
# Example 1 --------------------------------
trial_data <- trial |>
  dplyr::select(trt, age, marker, stage)

tbl_listing(trial_data, keys = c(trt, stage))

}
