% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_listing.R
\name{tbl_listing}
\alias{tbl_listing}
\title{Create listings from a data frame}
\usage{
tbl_listing(
  df,
  key_cols = NULL,
  sort_cols = NULL,
  add_blank_key_duplicates = TRUE,
  ...
)
}
\arguments{
\item{df}{(\code{data.frame})\cr
a data frame containing the data to be displayed in the listing.}

\item{key_cols}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to be used as key columns in the listing.}

\item{sort_cols}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to be used for sorting the data frame before creating the listing. It defaults to \code{key_cols}.}

\item{add_blank_key_duplicates}{(\code{logical})\cr
whether to add blank values where key columns have duplicate values. Defaults to \code{TRUE}.}

\item{...}{other objects that will be added to the \code{{gtsummary}} object list.}
}
\description{
Create listings from a data frame
}
\details{
This function converts a data frame into a listing format and it relies on the user's pre-processing
to decide which column to show and select unique or non-unique rows. Also sorting depends on the user
pre-processing.
}
\examples{
# Example 1 --------------------------------
trial_df <- trial |>
  dplyr::select(trt, age, marker, stage)

tbl_listing(trial_df, key_cols = c(trt, stage))

}
