% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_mmrm.R
\name{tbl_mmrm}
\alias{tbl_mmrm}
\alias{get_mmrm_results}
\title{Get and display MMRM Results in a Formatted Table}
\usage{
get_mmrm_results(fit_mmrm, arm, visit, conf_level = 0.95)

tbl_mmrm(
  mmrm_df,
  base_df = NULL,
  arm,
  visit,
  baseline_aval = NULL,
  digits = c(2, 3, 4)
)
}
\arguments{
\item{fit_mmrm}{(\code{mmrm} model object)\cr
A fitted MMRM model object, typically created using the \code{mmrm} function from the \code{mmrm} package. This object
should contain the necessary information to extract adjusted means, differences, confidence intervals, and
p-values for the specified visits and arms.}

\item{arm}{(\code{string})\cr
The column in \code{mmrm_df} and \code{base_df} that identifies the treatment arms.
This will be used to divide the results in columns. First value is reference.}

\item{visit}{(\code{string})\cr
The column in \code{mmrm_df} and \code{base_df} that identifies the visits.
This will be used to stratify the results in rows.}

\item{conf_level}{(\code{numeric})\cr
The confidence level to use when calculating confidence intervals for the adjusted means and differences.
Default is 0.95 for 95\% confidence intervals.}

\item{mmrm_df}{(\code{data.frame})\cr
A tidy data frame containing the MMRM results. This should include
columns for the visit, arm, adjusted means, differences, confidence intervals, and
p-values. The data frame should be structured in a way that allows for stratification by visit and arm.
Usually the output of \code{\link[=get_mmrm_results]{get_mmrm_results()}}.}

\item{base_df}{(\code{data.frame})\cr
A data frame containing baseline measurements. This should include columns for
the visit, arm, and baseline values. The function will summarize this data
to provide baseline statistics in the final table. If \code{base_df} is not provided
or does not contain any rows, the function will simply omit the baseline summary
section from the final table.}

\item{baseline_aval}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
The column in \code{base_df} that contains the baseline values to be summarized.
Unused if \code{base_df} is not provided or does not contain any rows.}

\item{digits}{(\code{numeric})\cr
A numeric vector of length 3 specifying the number of decimal places for: 1) Estimates/CIs, 2) Standard
Errors, and 3) P-values. Default is \code{c(2, 3, 4)}.}
}
\value{
A \code{data.frame} containing the estimated marginal means (adjusted means)
and contrasts (differences in adjusted means) for each visit and arm,
along with their standard errors, confidence intervals, degrees of freedom,
and sample sizes. This data frame is structured to facilitate the creation of a
formatted table using \code{\link[=tbl_mmrm]{tbl_mmrm()}}.

\code{tbl_mmrm} returns a 'gtsummary' table object.
}
\description{
These functions take a fitted MMRM model object and creates a formatted table,
following the style of the MMRM template. It combines baseline summary statistics
(if available) with the MMRM results, presenting them in a clear and organized manner.
}
\examples{
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true") && requireNamespace("mmrm", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(mmrm)
fv_dt <- mmrm::fev_data |>
  dplyr::mutate(
    ARMCD = sprintf(
      "\%s\n(N = \%d)", ARMCD,
      table(mmrm::fev_data$ARMCD)[ARMCD]
    ),
    ARMCD = factor(ARMCD)
  )
\dontshow{\}) # examplesIf}
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true") && requireNamespace("mmrm", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Fit an MMRM model using the FEV data
fit_mmrm <- mmrm::mmrm(
  # us -> unstructured cov structure
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  data = fv_dt
)
mmrm_results <- get_mmrm_results(fit_mmrm, arm = "ARMCD", visit = "AVISIT", conf_level = 0.95)
\dontshow{\}) # examplesIf}
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true") && requireNamespace("mmrm", quietly = TRUE)) withAutoprint(\{ # examplesIf}
tbl_mmrm(
  mmrm_results,
  fv_dt |> dplyr::mutate(AVISIT = "Baseline"),
  arm = "ARMCD", visit = "AVISIT", baseline_aval = "FEV1"
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gg_mmrm_lineplot]{gg_mmrm_lineplot()}} for visualizing MMRM results.
}
