% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_null_report.R
\name{tbl_null_report}
\alias{tbl_null_report}
\title{Creates null report}
\usage{
tbl_null_report(
  label = "No observations met the reporting criteria for this output."
)
}
\arguments{
\item{label}{(\code{string})\cr label to display in the header of the null report. It defaults to
"No observations met the reporting criteria for this output."}
}
\description{
This function creates a null report for tables or listings without any statistics.
}
\examples{
tbl_null_report(label = "No data available for the selected criteria.")

}
