% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_roche_subgroups.R
\name{tbl_roche_subgroups}
\alias{tbl_roche_subgroups}
\title{Subgroup Analyses}
\usage{
tbl_roche_subgroups(data, rsp, by, subgroups, .tbl_fun)
}
\arguments{
\item{data}{(\code{data.frame}, \code{survey.design})\cr
a data frame or survey object}

\item{rsp}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variable to use in responder rate calculations.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variable to make comparison between groups.}

\item{subgroups}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to perform stratified analyses for.}

\item{.tbl_fun}{(\code{function})
A function or formula. If a \emph{function}, it is used as is.
If a formula, e.g. \code{~ .x \%>\% tbl_summary() \%>\% add_p()}, it is converted to a function.
The stratified data frame is passed to this function.}
}
\value{
a 'gtsummary' table
}
\description{
Function adapted from \code{gtforester::tbl_subgroups()}.
}
\examples{
# prepare sample data
df_adtte <- data.frame(
  time = rexp(100, rate = 0.1),
  status = sample(c(0, 1), 100, replace = TRUE),
  arm = sample(c("Arm A", "Arm B"), 100, replace = TRUE),
  grade = sample(c("I", "II"), 100, replace = TRUE),
  strata = sample(c("1", "2"), 100, replace = TRUE)
) |>
  mutate(arm = relevel(factor(arm), ref = "Arm A")) # Set Reference

# logistic regression -------------------------------------------------------
df_adtte |>
  tbl_roche_subgroups(
    rsp = "status",
    by = "arm",
    subgroups = c("grade"),
    .tbl_fun =
      ~ glm(status ~ arm, data = .x) |>
        tbl_regression(
         show_single_row = arm,
         exponentiate = TRUE,
         tidy_fun = broom.helpers::tidy_parameters
       )
 ) |>
 modify_header(starts_with("estimate") ~ "Odds Ratio")

# coxph regression ----------------------------------------------------------
df_adtte |>
  tbl_roche_subgroups(
    rsp = status,
    by = arm,
    subgroups = c(grade, strata),
    ~ survival::coxph(
      survival::Surv(time, status) ~ arm,
      data = .x,
      ties = "exact"
    ) |> # Exact Ties
      tbl_regression(
        show_single_row = arm,
        exponentiate = TRUE # Get Hazard Ratios
      )
  ) |>
  modify_header(starts_with("estimate") ~ "Hazard Ratio")

}
