% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survfit_times.R
\name{tbl_survfit_times}
\alias{tbl_survfit_times}
\alias{add_overall.tbl_survfit_times}
\title{Survival Times}
\usage{
tbl_survfit_times(
  data,
  times,
  y = "survival::Surv(time = AVAL, event = 1 - CNSR, type = 'right', origin = 0)",
  by = NULL,
  label = "Time {time}",
  statistic = c("{n.risk}", "{estimate}\%", "{conf.low}\%, {conf.high}\%"),
  estimate_fun = label_style_number(digits = 1, scale = 100),
  method.args = list(conf.int = 0.95)
)

\method{add_overall}{tbl_survfit_times}(
  x,
  last = FALSE,
  col_label = "**Overall**  \\nN = {gtsummary::style_number(N)}",
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
A data frame}

\item{times}{(\code{numeric})\cr
a vector of times for which to return survival probabilities.}

\item{y}{(\code{string} or \code{expression})\cr
A string or expression with the survival outcome, e.g. \code{survival::Surv(time, status)}.
The default value is
\code{survival::Surv(time = AVAL, event = 1 - CNSR, type = "right", origin = 0)}.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}, which returns results for the unstratified model.}

\item{label}{(\code{string})\cr
Label to appear in the header row. Default is \code{"Time {time}"}, where
the glue syntax injects the time estimate into the label.}

\item{statistic}{(\code{character})\cr
Character vector of the statistics to report.
May use any of the following statistics:
\code{c(n.risk, estimate, std.error, conf.low, conf.high)},
Default is \code{c("{n.risk}", "{estimate}\%", "{conf.low}\%, {conf.high}\%")}}

\item{estimate_fun}{(\code{function}) \cr
Function used to style/round the \code{c(estimate, conf.low, conf.high)} statistics.}

\item{method.args}{(named \code{list})\cr
Named list of arguments that will be passed to \code{survival::survfit()}.

Note that this list may contain non-standard evaluation components, and
must be handled similarly to tidyselect inputs by using
rlang's embrace operator \code{{{ . }}} or \code{!!enquo()} when programming with this
function.}

\item{x}{(\code{tbl_survfit_times})\cr
A stratified 'tbl_survfit_times' object}

\item{last}{(scalar \code{logical})\cr
Logical indicator to display overall column last in table.
Default is \code{FALSE}, which will display overall column first.}

\item{col_label}{(\code{string})\cr
String indicating the column label. Default is \code{"**Overall**  \nN = {style_number(N)}"}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a gtsummary table
}
\description{
Create a gtsummary table with Kaplan-Meier estimated survival estimates
and specified times.
}
\examples{
# Example 1 ----------------------------------
tbl_survfit_times(
  data = cards::ADTTE,
  by = "TRTA",
  times = c(30, 60),
  label = "Day {time}"
) |>
  add_overall()
}
