% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_roche.R
\name{theme_gtsummary_roche}
\alias{theme_gtsummary_roche}
\title{Roche Theme}
\usage{
theme_gtsummary_roche(set_theme = TRUE)
}
\arguments{
\item{set_theme}{(scalar \code{logical})\cr
Logical indicating whether to set the theme. Default is \code{TRUE}.
When \code{FALSE} the named list of theme elements is returned invisibly}
}
\value{
theme list
}
\description{
A gtsummary theme for Roche tables
\itemize{
\item flextable- and gt-printed tables are styled with reduced padding and font size.
\item Uses \code{label_roche_pvalue()} as the default formatting function for all p-values.
\item Font size defaults are 8 points for all the table by the footers that are 7 points.
\item Border defaults to \code{flextable::fp_border_default(width = 0.5)}.
\item The \code{add_overall(col_label)} default value has been updated.
}
}
\examples{
theme_gtsummary_roche()

tbl_demographics(trial, by = trt, include = c(age, grade))

reset_gtsummary_theme()
}
